requirements.txt
streamlit
opencv-python-headless
Pillow
ultralytics
paddlepaddle
paddleocr
numpy<2.0
setuptools
